tetris playable

+++ src/chip/bios.rs
@@ -1,6 +1,6 @@
 use std::{fs::File, io::Read, str};
 
-pub const BIOS_MAPPED_ADDRESS: usize = 0xFF50;
+pub const BIOS_MAPPED_ADDRESS: u16 = 0xFF50;
 pub struct Bios {
     pub data: [u8; 0x100],
     pub mapped: bool,
